Make form controls smaller
For some reason, the large from control stylings provided in the
Boostrap SASS gem aren't matching up with the Bootstrap stylings. There
appears to be an extra selector that sets `height:auto` and this is
causing there to be too much space on the inputs.  This selector resets
is back to the default value from Bootstrap.