Refactor: Clean up depreciation warning
`block_unknown_urls` was throwing a depreciation warning from Capybara.
The format is a little bit unusual because Rubocop was complaining about
the block format and so I used a symbol to proc. If there are two config
params, that wouldn't be possible, but I went ahead and did it here to
stay aligned with the Rubocop recommendations.